pyproject.toml
[project]
name = "autogen-getting-started"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "autogen-agentchat (==0.4.2)",
    "autogen-ext (==0.4.2)",
    "asyncio (==3.4.3)",
    "tiktoken (==0.8.0)",
    "openai (==1.59.7)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
